Some minor cleanup and documentation fixes. Regenerated reference documentation.
Renamed:
  build_module.py -> build.py
  inline.py -> inlining.py
to avoid shadowing by functions with same name which are
imported in __init__.py. I recommend deleting the old
installation before reinstalling to get rid of the old files.

Changed version number to 0.9.5 and updated changelog.

The documentation still has some TODOs and FIXMEs,
but it should be good enough for a 0.9.5 release.